+++ MapsIndoors.framework/Headers/NSURL+QueryParser.h
@@ -1,0 +1,8 @@
+#import <Foundation/Foundation.h>
+
+@interface NSURL (QueryParser)
+
+- (nonnull NSDictionary*)queryDictionary;
+
+@end
+