+++ CHANGELOG.md
@@ -8,7 +8,6 @@ Changelog for MapsIndoors for iOS. This document structure is based on [Keep a C
 ### Added
 ### Fixed
 ### Changed
-- MPLocation properties are now read only
 ### Removed
 -->
 
@@ -17,6 +16,10 @@ Changelog for MapsIndoors for iOS. This document structure is based on [Keep a C
 - Support for external location data sources using `[MapsIndoors registerLocationSources:sources]`
 - New location service `MPLocationService` to replace `MPLocationsProvider`
 - Added building and venues to the search experience
+- 
+
+### Changed
+- MPLocation properties are now read only
 
 ## [2.1.5] 2018-11-06
 ### Fixed